// src/types.ts
export interface Dependent {
  id: string;
  name: string;
  relationship: 'Self' | 'Father' | 'Son' | 'Mother' | 'Daughter' | 'Spouse';
  age: number;
  dob: string;
  avatar: string;
  adherenceRate: number; // 0 - 100%
  activeMedsCount: number;
  urgentAlertsCount: number;
  primaryCondition: string;
  primaryDoctor: string;
  doctorPhone: string;
  emergencyContact: string;
  emergencyPhone: string;
  hipaaAuthorized: boolean;
  notes?: string;
}
// ...
export interface Prescription {
  id: string;
  rxNumber: string;
  dependentId: string;
  medicationName: string;
  genericName: string;
  brandEquivalent: string;
  strength: string;
  form: string;
  doctorName: string;
  doctorSpecialty: string;
  doctorNpi: string;
  clinicName: string;
  clinicAddress: string;
  clinicPhone: string;
  prescribedDate: string;
  expirationDate: string;
  refillsTotal: number;
  refillsRemaining: number;
  status: 'Active' | 'Pending Renewal' | 'Expiring Soon' | 'Transferred';
  sig: string; // dosage directions
  qtyPrescribed: number;
  daysSupply: number;
  genericSubstitutionPermitted: boolean;
  qrVerificationCode: string;
}